backend/requirements.txt
# Core Flask
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0

# Database
sqlalchemy==2.0.23

# Security
werkzeug==3.0.1

# AI Models
openai-whisper==20231117
torch==2.1.2
transformers==4.36.2
accelerate==0.25.0

# Payments
stripe==7.8.0

# Production Server
gunicorn==21.2.0

# Environment Variables
python-dotenv==1.0.0